+++ src/WebPortal.h
@@ -1,8 +1,0 @@
-#pragma once
-
-class WebPortal
-{
-public:
-    static void begin();
-    static void update();
-};